Change float to int for easier functional test

+++ plotting/matrixMul/v2/tmp_var/mmul.cu
@@ -6,7 +6,7 @@
 // Matrix Multiplication kernel
 // Optimizations:
 //  Accumulate partial results in a temporary variable
-__global__ void matrixMul(float *a, float *b, float *c, int N){
+__global__ void matrixMul(int *a, int *b, int *c, int N){
     // Calculate the row and column for each thread
     int row = blockIdx.y * blockDim.y + threadIdx.y; 
     int col = blockIdx.x * blockDim.x + threadIdx.x;
@@ -25,7 +25,7 @@ __global__ void matrixMul(float *a, float *b, float *c, int N){
 }
 
 // Initialize a matrix with random numbers
-void init_matrix(float *m, int N){
+void init_matrix(int *m, int N){
     for(int i = 0; i < N * N; i++){
         m[i] = rand() % 100;
     }
@@ -34,15 +34,15 @@ void init_matrix(float *m, int N){
 int main(){
     // Problem size
     int N = 1 << 14;
-    size_t bytes = N * N * sizeof(float);
+    size_t bytes = N * N * sizeof(int);
 
     // Allocate host memory
-    float *h_a = (float*)malloc(bytes);
-    float *h_b = (float*)malloc(bytes);
-    float *h_c = (float*)malloc(bytes);
+    int *h_a = (int*)malloc(bytes);
+    int *h_b = (int*)malloc(bytes);
+    int *h_c = (int*)malloc(bytes);
     
     // Allocate device memory
-    float *d_a, *d_b, *d_c;
+    int *d_a, *d_b, *d_c;
     cudaMalloc(&d_a, bytes);
     cudaMalloc(&d_b, bytes);
     cudaMalloc(&d_c, bytes);
